+++ example/src/main/kotlin/example/backends/OnHttp4k.kt
@@ -51,6 +51,8 @@ val http4kRoutes: List<ServerEndpoint> = listOf(
 
     strict handledNow { (term, key, jar) -> Strictly(term, key, jar) },
 
+    roundtrip handledNow { (inPath, inQuery) -> RoundTrip(inPath, inQuery) },
+
     motd handledNow { "Be excellent to each other." },
 
     forget handledWith { _ -> },