+++ example/src/main/kotlin/example/backends/OnPekko.kt
@@ -50,6 +50,8 @@ val pekkoRoutes: List<ServerEndpoint> = listOf(
 
     strict handledNow { (term, key, jar) -> Strictly(term, key, jar) },
 
+    roundtrip handledNow { (inPath, inQuery) -> RoundTrip(inPath, inQuery) },
+
     motd handledNow { "Be excellent to each other." },
 
     forget handledWith { _ -> },